another fix to Bug 812 - ERROR 22003: The resulting value is outside the range for the data type INTEGER.

+++ java/engine/org/apache/derby/impl/sql/execute/CountAggregator.java
@@ -66,7 +66,7 @@ public final class CountAggregator
 	}
 
 	public void add(DataValueDescriptor addend) throws StandardException{
-		value+=addend.getInt();
+		value+=addend.getLong();
 	}
 
 	/**
